+++ symbolic_trace/opcode_translator/executor/opcode_executor.py
@@ -14,7 +14,7 @@ from ...utils import (
 )
 from ..instruction_utils import get_instructions
 from .function_graph import FunctionGraph
-from .source import ConstTracker, DummyTracker, GlobalTracker, LocalTracker
+from .tracker import ConstTracker, DummyTracker, GlobalTracker, LocalTracker
 from .variables import (
     Guard,
     ListVariable,
