+++ symbolic_trace/opcode_translator/executor/tracker.py
